Fix single select behavior
Prompt:
```
When using the Filters component with `isMulti=false` (or when there is only one facet supplied) I'd like the options in the dropdown to continue to include all options for that facet, like a Select.

Currently, selecting an option removes any options from the menu that can't be selected because the current group of `items` doesn't have those properties; can we rework this behavior so that I can still see a list of composers when I have a composer selected via isMulti=false and let the user select a different composer without first clearing the existing composer?

I'd like the behavior on the situations where there are multiple facets to remain as it currently is, where the facets that have 0 possible values are not included in the dropdown / available options
```