Music Library: Fix bugs with Filters component (#55)
* Fix counts on Filters component

(Prompt below)

---
I found a minor bug with the Filters component

In the `formatOptionLabel` function, the `count` needs to be configurable to reflect counts of different objects

On the "Works" page in the IndexPage component, we are showing a list of works; when I use the Filters on that page, the facets are Composers (which is correct) but the counts are concerts, which is incorrect. The counts should be number of works on the works page, to reflect "this composer has this many works"

---

* Show facet label on selected options in Filters

Prompt:

```
I'd like it if the "selected options" that show in the Select when the dropdown is closed would have a prefix with the type of object they are; for example, when you click "2024-2025" under "Season" in the dropdown, I'd like the value to show "Season: 2024-2025" in the Filters when the dropdown is dismissed
```

then
```
Can we only make this label show when `isMulti` is true?
```